Fix a bug caused by python3's idea of integer-dividion remainder.
git-svn-id: svn://svn.code.sf.net/p/wsjt/wsjt/trunk@4336 ab8295b8-cf94-4d9e-aec4-7959e3be5d79

+++ wsjt.py
@@ -300,7 +300,7 @@ def dbl_click_call(t,t1,rpt,event):
             nsec=60*int(t1[i3+2:i3+4]) + int(t1[i3+4:i3+6])
         except:
             nsec=0
-        if setseq.get(): TxFirst.set((nsec/Audio.gcom1.trperiod)%2)
+        if setseq.get(): TxFirst.set(int((nsec/Audio.gcom1.trperiod)%2))
         lookup()
         GenStdMsgs()
         if (mode.get()[:4]=='JT65' or \
